package done

+++ app/Http/Controllers/PackageController.php
@@ -9,6 +9,7 @@ use App\Models\Package;
 use App\Http\Requests\StorePackageRequest;
 use App\Http\Requests\UpdatePackageRequest;
 use App\Repositories\interfaces\PackageRepositoryInterface;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Response;
 
 class PackageController extends Controller
@@ -28,9 +29,11 @@ class PackageController extends Controller
      *
      * @param IndexPackageRequest $request
      * @return PackageResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexPackageRequest $request)
     {
+        $this->authorize('viewAny', Package::class);
         $items = $this->packageRepository->findBy($request->all());
         return new PackageResourceCollection($items);
     }
@@ -40,9 +43,11 @@ class PackageController extends Controller
      *
      * @param StorePackageRequest $request
      * @return PackageResource
+     * @throws AuthorizationException
      */
     public function store(StorePackageRequest $request)
     {
+        $this->authorize('create', Package::class);
         $item = $this->packageRepository->save($request->all());
         return new PackageResource($item);
     }
@@ -52,9 +57,11 @@ class PackageController extends Controller
      *
      * @param Package $package
      * @return PackageResource
+     * @throws AuthorizationException
      */
     public function show(Package $package)
     {
+        $this->authorize('view', Package::class);
         $item = $this->packageRepository->findOne($package->id);
         return new PackageResource($item);
     }
@@ -65,9 +72,11 @@ class PackageController extends Controller
      * @param UpdatePackageRequest $request
      * @param Package $package
      * @return PackageResource
+     * @throws AuthorizationException
      */
     public function update(UpdatePackageRequest $request, Package $package)
     {
+        $this->authorize('update', Package::class);
         $item = $this->packageRepository->update($package, $request->all());
         return new PackageResource($item);
     }
@@ -77,9 +86,11 @@ class PackageController extends Controller
      *
      * @param Package $package
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(Package $package)
     {
+        $this->authorize('delete', Package::class);
         $this->packageRepository->delete($package);
     }
 }
